+++ R/package.R
@@ -2,36 +2,23 @@
 #'
 #' @description
 #' Alternating optimization of (high-dimensional) functions is an iterative
-#' procedure for minimizing (or maximizing) jointly over all parameters by
-#' alternately optimizing parameter subsets.
+#' procedure for optimizing jointly over all parameters by alternately
+#' optimizing parameter subsets.
 #'
 #' @docType package
+#'
 #' @name ao
+#'
 #' @keywords
 #' internal
+#'
+#' @import optimizeR
 "_PACKAGE"
 
 ## usethis namespace: start
 #' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL
-
-#' @noRd
-#' @keywords
-#' internal
-
-.onAttach <- function(lib, pkg) {
-  msg <- c(
-    paste0(
-      "Thanks for using {ao} ", utils::packageVersion("ao")
-    ),
-    ", happy alternating optimization!\n",
-    "See https://loelschlaeger.github.io/ao for help.\n",
-    "Type 'citation(\"ao\")' for citing this R package."
-  )
-  packageStartupMessage(msg)
-  invisible()
-}
 
 #' @noRd
 #' @keywords