+++ tests/testthat/test-utils.R
@@ -1,5 +1,0 @@
-test_that("euclidean works", {
-  x <- 1:10
-  y <- 1:10
-  expect_equal(euclidean(x, y), 0)
-})